Remove code for external state graph visualiser
Closes #758
