+++ FstPlugin/src/org/workcraft/plugins/fst/tasks/WriteSgConversionTask.java
@@ -15,7 +15,7 @@ import org.workcraft.interop.Exporter;
 import org.workcraft.plugins.fst.Fst;
 import org.workcraft.plugins.fst.interop.SgImporter;
 import org.workcraft.plugins.petri.PetriNetModel;
-import org.workcraft.plugins.petrify.tasks.WriteSgTask;
+import org.workcraft.plugins.petrify.WriteSgTask;
 import org.workcraft.plugins.shared.tasks.ExternalProcessResult;
 import org.workcraft.plugins.stg.interop.StgFormat;
 import org.workcraft.tasks.ProgressMonitor;
