+++ PetrifyExtraPlugin/src/org/workcraft/plugins/petrify/WriteSgTask.java
@@ -1,25 +1,32 @@
-package org.workcraft.plugins.petrify.tasks;
+package org.workcraft.plugins.petrify;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.workcraft.plugins.petrify.PetrifySettings;
+import org.workcraft.interop.ExternalProcessListener;
 import org.workcraft.plugins.shared.tasks.ExternalProcessResult;
 import org.workcraft.plugins.shared.tasks.ExternalProcessTask;
 import org.workcraft.tasks.ProgressMonitor;
 import org.workcraft.tasks.Result;
 import org.workcraft.tasks.Result.Outcome;
 import org.workcraft.tasks.Task;
+import org.workcraft.util.DataAccumulator;
 import org.workcraft.util.ToolUtils;
 
-public class DrawAstgTask implements Task<ExternalProcessResult> {
+public class WriteSgTask implements Task<ExternalProcessResult>, ExternalProcessListener {
     private final List<String> options;
     private final File inputFile;
     private final File outputFile;
     private final File workingDirectory;
 
-    public DrawAstgTask(List<String> options, File inputFile, File outputFile, File workingDirectory) {
+    private ProgressMonitor<? super ExternalProcessResult> monitor;
+
+    private final DataAccumulator stdoutAccum = new DataAccumulator();
+    private final DataAccumulator stderrAccum = new DataAccumulator();
+
+    public WriteSgTask(List<String> options, File inputFile, File outputFile, File workingDirectory) {
         this.options = options;
         this.inputFile = inputFile;
         this.outputFile = outputFile;
@@ -28,16 +35,19 @@ public class DrawAstgTask implements Task<ExternalProcessResult> {
 
     @Override
     public Result<? extends ExternalProcessResult> run(ProgressMonitor<? super ExternalProcessResult> monitor) {
+        this.monitor = monitor;
         ArrayList<String> command = new ArrayList<>();
 
         // Name of the executable
         String toolName = ToolUtils.getAbsoluteCommandPath(PetrifySettings.getCommand());
         command.add(toolName);
-        command.add("-draw_astg");
+        command.add("-write_sg");
 
         // Built-in arguments
-        for (String arg : options) {
-            command.add(arg);
+        if (options != null) {
+            for (String arg : options) {
+                command.add(arg);
+            }
         }
 
         // Input file
@@ -53,7 +63,6 @@ public class DrawAstgTask implements Task<ExternalProcessResult> {
 
         ExternalProcessTask task = new ExternalProcessTask(command, workingDirectory);
         Result<? extends ExternalProcessResult> res = task.run(monitor);
-
         if (res.getOutcome() != Outcome.SUCCESS) {
             return res;
         }
@@ -65,4 +74,29 @@ public class DrawAstgTask implements Task<ExternalProcessResult> {
             return Result.failure(retVal);
         }
     }
+
+    @Override
+    public void errorData(byte[] data) {
+        try {
+            stderrAccum.write(data);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        monitor.stderr(data);
+    }
+
+    @Override
+    public void outputData(byte[] data) {
+        try {
+            stdoutAccum.write(data);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        monitor.stdout(data);
+    }
+
+    @Override
+    public void processFinished(int returnCode) {
+    }
+
 }
